apps/stream/pyproject.toml
[tool.poetry]
name = "traffic-stream"
version = "0.1.0"
description = ""
authors = ["2uanDM <hokage321xxx@gmail.com>"]
readme = "README.md"
packages = [
  {include = "src"}
]

[tool.poetry.dependencies]
python = "^3.11"
kafka-python = "^2.0.2"
polars = "^1.17.1"
clickhouse-connect = "^0.8.10"
geopandas = "^1.0.1"
python-dotenv = "^1.0.1"
pydantic = "^2.10.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
